Rewrote the way the core handles context's, seems to be stable. Script::FlushCache() is called before a script is compiled, this is a workaround to a bug I found. Events are working, I tested with eventtest.dbj If you dont clear the flush before reloading a script strange things happens. Fixed some bugs within BuildObject. Tested with evilcode.dbj - seemed to work fine Tested with dual evilcode.dbj -
[Mar/19/2009 - 11:57:29] [Error] Code (137) default2.dbj/line 12: out of memory
Line: (null)
[Mar/19/2009 - 11:57:29] [Error] Code (137) default.dbj/line 13: out of memory
Line: (null)

ran out of memory, however this is due to the design of the script, not a flaw in the core (i think)
.stop still seems kind of buggy, i tried tweaking it around however, im not getting much success.

unitbug.dbj - seemed to work decently, it crashed after a few minutes though - pure virtual call, screenhooks still a bit buggy for intensive use.


i hope this runs other scripts with success.