Re-added NTAbaddon.dbl from sirgijoe (thanks sirgijoe!)
Addition made to universal config file:
NTConfig_ClearAbaddon = true;			//	Set to true to clear Abaddon of monsters (NOTE: must be running 'Abaddon' in your script array)

When running NTAbaddon.dbl this will set whether or not to clear all of Abaddon.  (Note: This is optional! You have the choice to pop the chest only)
If NTConfig_OpenChests is set to true, bot will open the super-chest in Abaddon.